# fct-frontend/api-spec/openapi.yaml
openapi: 3.1.1
info:
    title: REST API for the website of "Academia Fabiana Perez"
    description: REST API for the website of "Academia Fabiana Perez"
    version: 1.0.0
servers:
    - url: http://localhost:8080/api/v1
paths:
...
    /product-categories:
        get:
            summary: Get all product categories
            operationId: getAllProductCategories
            tags:
                - Product Categories
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/ProductCategory"
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        post:
            summary: Create a new product category
            operationId: createProductCategory
            tags:
                - Product Categories
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/AddProductCategoryRequest"
            responses:
                201:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ProductCategory"
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                409:
                    description: Product category already exists
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
...
    /product-categories/{id}:
        get:
            summary: Get a product category by id
            operationId: getProductCategoryById
            tags:
                - Product Categories
            parameters:
                - name: id
                  in: path
                  description: Product category id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ProductCategory"
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Product category not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        put:
            summary: Update a product category
            operationId: update productCategory
            tags:
                - Product Categories
            parameters:
                - name: id
                  in: path
                  description: Product category id
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                description: Product category to update
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateProductCategoryRequest"
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ProductCategory"
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Product category not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        delete:
            summary: Delete a product category
            operationId: deleteProductCategory
            tags:
                - Product Categories
            parameters:
                - name: id
                  in: path
                  description: Product category id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                204:
                    description: Successful operation
                400:
                    description: Invalid status value
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Product category not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
...
components:
    schemas:
...
        ProductCategory:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                vatPercentage:
                    type: number
                    format: double
            required: [id, name, vatPercentage]
...
        AddProductCategoryRequest:
            type: object
            properties:
                name:
                    type: string
                    x-field-extra-annotation: '@NotBlank(message = "Name field cannot be blank")'
                vatPercentage:
                    type: number
                    format: double
                    minimum: 0
                    maximum: 1
            required: [name, vatPercentage]
        UpdateProductCategoryRequest:
            type: object
            properties:
                name:
                    type: string
                    x-field-extra-annotation: '@NotEmpty(message = "Name field cannot be empty")'
                vatPercentage:
                    type: number
                    format: double
                    minimum: 0
                    maximum: 1
            required: [name, vatPercentage]

        ## ---- Invoice ----
...
        ErrorMessage:
            type: object
            properties:
                timestamp:
                    type: string
                    format: date-time
                    example: "1970-01-01T00:00+18:00"
                status:
                    type: integer
                    format: int32
                    example: 404
                error:
                    type: string
                    example: "Not Found"
                message:
                    type: string
                    example: "Resource not found"
                path:
                    type: string
                    example: "/api/v1/resource/1"
            required: [timestamp, status, error, message, path]
...
    securitySchemes:
        BearerAuth:
            description: Bearer token using a JWT
            type: http
            scheme: bearer
            bearerFormat: JWT